🌱 Remove omitempty from struct fields tags in API scaffolding
The `omitempty` JSON tag has been ineffective for nested structs, was replaced by `omitzero` in go 1.24 and is now flagged by the modernize linter. It was still used as a fallback by controller-gen to determine if a field was required in the absence of explicit marker, but since scaffolded API types now include explicit `+optional` and `+required` markers, the tag is can be removed.

I also aligned ListMeta tags with ObjectMeta, and a file in docs that was forgotten.